working baseline

+++ bin/infra-init.sh
@@ -75,10 +75,9 @@ fi
 #modprobe autofs4
 
 if [[ -e /dev/disk/by-label/EFI ]]; then
-  mkdir -p /run/media/efi /boot
+  mkdir -p /run/media/efi
   mount -o ro,noexec,nosuid,nodev /dev/disk/by-label/EFI /run/media/efi
-  mount --bind /run/media/efi /boot
-#  mount /run/media/efi/kernel/modules /usr/lib/modules
+  mount --bind /run/media/efi $NEWROOT/boot
   mp="/run/media/efi"
 
 rm -rf $NEWROOT/usr/lib/modules /usr/lib/modules
@@ -89,18 +88,22 @@ ln -sf $NEWROOT/usr/lib/modules /usr/lib/
  modprobe autofs4
 fi
 
-#if [[ -e /dev/disk/by-label/home ]]; then
+if [[ -e /dev/disk/by-label/home ]]; then
   mkdir -p /home
   echo "LABEL=home /home auto noauto,x-systemd.automount,x-systemd.idle-timeout=5min 0 2" >> $R/etc/fstab
   ln -sf /home $R/Users
 #  mount /dev/disk/by-label/home /home
-#fi
+fi
 
-#if [[ -e /run/initramfs/live ]]; then
-#  mount /run/initramfs/live/kernel/modules /usr/lib/modules
-#  mkdir -p /boot
-#  mount --bind /run/initramfs/live /boot
-#fi
+if [[ -e /run/initramfs/live ]]; then
+rm -rf $NEWROOT/usr/lib/modules /usr/lib/modules
+mkdir -p $NEWROOT/usr/lib/modules
+  mount /run/initramfs/live/kernel/modules $NEWROOT/usr/lib/modules
+  mkdir -p $NEWROOT/boot
+  mount --bind /run/initramfs/live $NEWROOT/boot
+
+  modprobe autofs4
+fi
 
 # --- detect the environment
